ocpp-server/requirements.txt
websockets>=10.0
ocpp>=0.26.0
paho-mqtt>=1.6.0
